Implement the 1-variant fallback in the image mutator
Some of the modules that deal with different image variants are not
straightforward to replace with equivalent properties. Just reimplement
the 1-variant fallback in the image mutator so that we don't have
to worry about it for the moment.

Having only the image mutator and not the arch/os mutators implement
this should prevent the "more fallbacks than before" issues that I saw
when converting all 3 of them to have this behavior.

Bug: 372091092
Test: m nothing
Change-Id: I9588e0bb9436ac5ddad750c1e70d883d5fe07630