Move disabling rust sanitizers for linux_musl_x86 earlier
Setting Never at the end of begin() is too late, and SanitizerEnabled
is still being set.  This causes the extra sanitizer versions to still
be created, but HideFromMake doesn't get called, causing conflicts
in the generated Android-${TARGET_PRODUCT}.mk.

Set Never earlier so that the rest of setting up sanitizers is skipped.

Fixes: 376075440
Flag: EXEMPT bugfix
Test: m USE_HOST_MUSL=true BUILD_HOST_static=true SANITIZE_HOST=address ASAN_OPTIONS="detect_leaks=0" ${OUT_DIR}/host/linux-x86/bin/dex2oat64
Change-Id: I48fc1f6fd9051b51c34374c56edacc30fb585642