Stop dumping/diffing preview ABIs.
Tracking the unfinalized API levels is causing more harm than good.
It's a major source of merge conflicts and the codenames themselves
are subject to change. It's also a pretty slow extra step for API
authors.

We can bring this back if we ever manage to get all the API surfaces
onto STG, because then we could use the STG files for signaling API
review instead of the *.map.txt filter, but since the others currently
don't use STG and don't track preview API, we can't do that today.

Bug: None
Test: treehugger
Change-Id: I309bfeef88c4f2c8a9112beddef8bfe88da9ef32