Support override_android_(app|apex) in deps of android_filesystem
These modules implicitly override their base module. e.g. if
com.android.foo and com.company.android.foo are listed in deps, only the
latter should be installed.

Override modules are implemented as an "override" variant of the base
module, and all the build actions are generated there. To not install
the overridden app/apex, the following properties are updated in the
_overriding_ PackagingSpec variant of the base apex
1. overrides: base apex (e.g. com.android.foo)
2. owner: overriding apex (e.g. com.company.android.foo)

Test: go test ./filesystem
Change-Id: I59f17134dc6ed99a252977c3615b1be5d6d746ba