1:1 mapping between prebuilt_kernel_modules and _dlkm filesystem modules
This is a correction to the autogeneration logic introduced in
https://r.android.com/3329416, which would create multiple
prebuilt_kernel_modules for a specific partition if the .ko files are
distributed in the tree. In addition to assembling and installing the
.ko files, `prebuilt_kernel_modules` are also responsible for creating
dependency metadata files by running `depmod`. If we autogenerate
multiple prebuilt_kernel_modules for a specific partition, the
dependency metadata files will be incorrect, and we will also have
multiple installation rules for /lib/modules/modules.(load|dep|...)

To prevent introducing complexity of intermediate filegroups, the
`prebuilt_kernel_modules` will be generated in the root directory for
now.

Test: m nothing
Bug: 377562851
Change-Id: Ic5d0c95f2bf32a9a97ba0489801fe2bfd4928aef