Don't use DirectlyInAnyApex to enable dex and coverage
DirectlyInAnyApex was being used to automatically enable dexing
and to enable coverage.  There is only one module in an apex that
isn't already being dexed, so set compile_dex in that module and
remove the automatic dexing.  Coverage is only useful on device
modules if they are dexed, so convert the shouldInstrumentInApex
test to check if it is dexed instead.

This reomves the last usages of DirectlyInAnyApex from the java
package.

Bug: 372543712
Test: No change to build.ninja
Change-Id: I1d1738305ed97d85669945def75e9916c2b41f3c