Flush verbose log every second
The gzipped verbose log can buffer for an arbitrarily long time,
which could lose logs in case of a crash, and is annoying when
trying to see the ouptut while the build is still running.  Move
the writes to the verbose log to a goroutine that also calls
gzip.Writer.Flush() every second to make sure the file on disk
is as up to date as possible.

Test: manual
Change-Id: Ib66dad85bbf409bc8327fe11fd78667952de9138