Add missing root dirs/symlinks to soong-built system image
These root dirs and symlinks are normally set up in a rather hardcoded
fashion in create_root_structure.mk. In the handwritten soong-built
images, they are listed out explicitly, so do the same thing in
the autogenerated images.

The file list diff test didn't catch them because the soong file
list only considers stuff under the rebased_dir, and in make the root
dir is passed to build_image.py separately and only the system staging
dir is considered for the file list.

Bug: 381120092
Test: Build make/soong system images, mount, diff root directory
Change-Id: I5f4919b92715369c54ca25cfe4815175825650c6