Add selinux contexts to autogenerated partitions
Filesystems had a selinux_contexts property that then compiled the
selinux contexts. But in make, it uses the result of the
file_contexts_bin_gen module, which already runs the compilation step.
Add a precompiled_file_contexts property to accept that compiled
file.

Also add 2 missing symlinks to the system partition.

Bug: 381120092
Test: diff'd make and soong build_image prop files
Change-Id: Ia0681fa4afe43675b730eaf0857dc6b15771534e