Use partition intead of "root" to assemble filesystem artifacts
The artifacts for filesystem packaging are currently assembled in
```
$intermediates/$module/android_common/root/$base_dir
```

This CL changes this to
```
$intermediates/$module/android_common/$partition/$base_dir
e.g.
$intermediates/$module/android_common/system/system (system)
$intermediates/$module/android_common/system_ext (system_ext)
```

The motivatiton for this change is to fix diffs in
BuildManifestSystemExt.apk. The build_manifest.pb of this apk is
generated from the base of $PRODUCT_OUT. This means that the paths of
the artifacts contains the partition prefix. diff https://diff.googleplex.com/#key=9xuSx7nvP251

An alternative would be to use partition as the `Base_dir`, but this
would create an additional top-level directory in the soong img files
(make img files do not have a top-level partition directory for non
system images)

Test: 383144733
Bug: Built and mounted make and soong system_ext.img for AOSP CP. NOTICE
is the last diff

Change-Id: Iee53113a7e00dcda9d987f2aabf9605740bbe7fe