Reapply "Remove unassigned event log tags feature"
.logtags files are a mapping of an integer -> name. However, there
is an unused feature where you could put a '?' instead of an integer,
and then the build system would assign a number to it, taking care
to avoid all other already-assigned numbers across the source tree.

This feature makes it hard to support some things we're working towards,
like incremental soong and treble common system images, so remove it.

Also modernize the python scripts a little.

The additions since the revert are the cleanups from aosp/3403019, and
removing a missed reference to all-event-log-tags.txt in java.mk.

This reverts commit 692c4e04990bfb94fdc1f0c3853f04ade7fe2757.

Bug: 382515940
Test: m event-log-tags
Change-Id: I8f08f3286f1275083e2074948f2feb39219f3687