Add the ability to always run errorprone inline
Currently, if you enable RUN_ERROR_PRONE=true, we create a seperate
jar from the main build which runs errorprone. But errorprone builds
also produce a regular built jar as a side effect, so you could use
it as your main java builder action. We don't do that because if you
toggled errorprone on/off, you'd invalidate all the java builds.

However, on CI, we always run with errorprone on, so there, it makes
more sense to always do the inline builds so that you don't have to
build everything twice.

Add a new environment variable to control running errorprone inline
or not.

Bug: 383626679
Test: Presubmits
Change-Id: I262af140fdeacb8342fe801f0c86f516d8fca587