Add AAPT flags to optimize the set of resources in the RRO apk.
The solution is based on adding flags in build/soong/java/app.go.
It will reduce the amount of resources in RRO. Also, it solves the problem of extra locales
 in the Setting and TvSettings applications (see AssetManager2.h, GetResourceLocales()).

Add 3 flags:
1. "--product" - specifies a list of product names to keep.
2. "-c" - provides a list of configurations.
For example, if you have dependencies on the support library, which contains translations for multiple languages, you can filter resources just for the given language configuration, like English or Spanish.
3. "--preferred-density" - allows AAPT2 to select the closest matching density and strip out all others.

Change-Id: I0da8bc05f349553188264824bfa5b53c27882ee3
Signed-off-by: Artem Anashkin <asanashkin@salutedevices.com>