Generate block_list only when building img for target_files.zip
The same block_list file is currently declared in the propfile of
primary .img file as well the secondary .img file (for
target_files.zip). Both the .img files are created in the same ninja
command, but in the current implementation the block_list file is
generated twice with the same contents.

This CL moves the block_list file to the image prop file of the
secondary .img creation.

Test: presubmits
Change-Id: I0f8f4e4c7bec1ae5e370ac97d6608f9754e719e3